fix: OSError - can't find edu tree during parsing some documents - other error do that, but I haven't time to look for this error

+++ run.py
@@ -86,11 +86,12 @@ def edu_parsing_multiprocess(parser, docs_id_range, edu_trees_dir,
 
                 processed += 1
         # skip documents that parsing returns errors
-        except (ValueError, IndexError, ZeroDivisionError) as err:
+        except (ValueError, IndexError, ZeroDivisionError, OSError) as err:
             logging.error(
                 'Error for doc #{}: {}. It has been skipped'.format(
                     document_id, str(err)))
-            shutil.rmtree(edu_tree_path)
+            if exists(edu_tree_path):
+                shutil.rmtree(edu_tree_path)
             errors += 1
 
         logging.info(
